chore: Move prisma back to node modules folder
