+++ backend/prisma/case_management/case_management.prisma
@@ -1,6 +1,6 @@
 generator client {
   provider        = "prisma-client-js"
-  output          = "../../prisma/case_management/generated"
+  output          = "../../node_modules/.prisma/case_management"
   previewFeatures = ["prismaSchemaFolder"]
   binaryTargets   = ["native", "debian-openssl-1.1.x"]
 }
