+++ docker-compose.yml
@@ -48,7 +48,7 @@ services:
 
   backend:
     container_name: cm-backend
-    entrypoint: sh -c "npm i && npm run start:dev"
+    entrypoint: sh -c "npm i && npm run prisma-generate-cm && npm run prisma-generate-inv && npm run start:dev"
     environment:
       <<: *postgres-vars
       NODE_ENV: development
